// src/types/schemas/stageDisplayState.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface StageDisplayState {
	mode: 'PROGRAM' | 'PREVIEW';
	showPlacementHelpers: boolean;
	message: {
		visible: boolean;
		mode: 'QUICK' | 'GENTLE';
		color: 'YELLOW' | 'RED' | 'GRAY';
		text: string;
	};
}